requirements.txt
langchain==0.1.17
langchain-community==0.0.26

sentence-transformers==2.6.1
transformers==4.36.2
faiss-cpu==1.7.4
huggingface-hub==0.16.4
accelerate==0.23.0

pdfplumber==0.9.0
camelot-py[cv]==0.10.1
tabula-py==2.7.0

python-slugify==8.0.1

fpdf2==2.7.4

python-dotenv==1.1.0

numpy==1.23.4
pandas==1.3.5
scikit-learn==1.3.2
nltk==3.8.1

tensorflow==2.13.0

tqdm==4.66.2
typing-extensions==4.10.0
openai==0.27.0
